Game server now exclusively uses the local horse data, fixed bug that broke server cancellation
