+++ src/game/horse/localHorses.ts
@@ -1,5 +1,10 @@
 import { Horse, IHorse } from "../../models/Horse.js";
+import { randomIndicesNoReplacement } from "../../random/random.js";
+import { Race } from "../race.js";
 import { InternalHorse } from "./horse.js";
+
+export const HORSE_POPULATION = 100
+export const HORSES_PER_RACE = 4
 
 /** Collection of horses in memory for the game server to access and manipulate */
 export let localHorses: InternalHorse[] = []
@@ -21,6 +26,15 @@ export function generateLocalHorsesFromSpecs(horseSpecs: IHorse[]) {
     console.log('Recreating local horse collection')
     localHorses = horseSpecs.map((hs) => new InternalHorse(hs))
     console.log('Finished recreating local horse collection')
+}
+
+export function createRace(): Race | null {
+    console.log('creating race')
+    let indices = randomIndicesNoReplacement(HORSE_POPULATION, HORSES_PER_RACE)
+    if (indices === null) { return null }
+    const race = new Race(indices.map((i) => localHorses[i]))
+    console.log('created race')
+    return race
 }
 
 await loadHorsesFromDatabase()
